OSIS-4671 : déplacement des fonctions instanciate_teacher_object et instanciate_attribution dans le repository
